+++ pipelined/bevy_pbr2/src/lib.rs
@@ -54,27 +54,27 @@ impl Plugin for PbrPlugin {
             .init_resource::<AmbientLight>()
             .add_system_to_stage(
                 CoreStage::PostUpdate,
-                render::update_directional_light_frusta
-                    .label(LightSystems::UpdateDirectionalLightFrusta)
+                update_directional_light_frusta
+                    .label(SimulationLightSystems::UpdateDirectionalLightFrusta)
                     .after(TransformSystem::TransformPropagate),
             )
             .add_system_to_stage(
                 CoreStage::PostUpdate,
-                render::update_point_light_frusta
-                    .label(LightSystems::UpdatePointLightFrusta)
+                update_point_light_frusta
+                    .label(SimulationLightSystems::UpdatePointLightFrusta)
                     .after(TransformSystem::TransformPropagate),
             )
             .add_system_to_stage(
                 CoreStage::PostUpdate,
-                render::check_light_visibility
-                    .label(LightSystems::CheckLightVisibility)
+                check_light_visibility
+                    .label(SimulationLightSystems::CheckLightVisibility)
                     .after(TransformSystem::TransformPropagate)
                     .after(VisibilitySystems::CalculateBounds)
-                    .after(LightSystems::UpdateDirectionalLightFrusta)
-                    .after(LightSystems::UpdatePointLightFrusta)
+                    .after(SimulationLightSystems::UpdateDirectionalLightFrusta)
+                    .after(SimulationLightSystems::UpdatePointLightFrusta)
                     // NOTE: This MUST be scheduled AFTER the core renderer visibility check
-                    //       because that resets entity ComputedVisibility for the first view
-                    //       which would override any results from this otherwise
+                    // because that resets entity ComputedVisibility for the first view
+                    // which would override any results from this otherwise
                     .after(VisibilitySystems::CheckVisibility),
             );
 
@@ -83,7 +83,7 @@ impl Plugin for PbrPlugin {
             .add_system_to_stage(RenderStage::Extract, render::extract_meshes)
             .add_system_to_stage(
                 RenderStage::Extract,
-                render::extract_lights.label(LightSystems::ExtractLights),
+                render::extract_lights.label(RenderLightSystems::ExtractLights),
             )
             .add_system_to_stage(
                 RenderStage::Prepare,
@@ -91,12 +91,12 @@ impl Plugin for PbrPlugin {
                 // _before_ the `prepare_views()` system is run. ideally this becomes a normal system when "stageless" features come out
                 render::prepare_lights
                     .exclusive_system()
-                    .label(LightSystems::PrepareLights),
+                    .label(RenderLightSystems::PrepareLights),
             )
             .add_system_to_stage(RenderStage::Queue, render::queue_meshes)
             .add_system_to_stage(
                 RenderStage::Queue,
-                render::queue_shadows.label(LightSystems::QueueShadows),
+                render::queue_shadows.label(RenderLightSystems::QueueShadows),
             )
             .add_system_to_stage(RenderStage::Queue, render::queue_shadow_view_bind_group)
             .add_system_to_stage(RenderStage::Queue, render::queue_transform_bind_group)
